19: Moving BulletedList component to globals
