--- src/components/primitives/BulletedList.tsx
@@ -1,14 +1,0 @@
-import styled from 'styled-components';
-
-export const BulletedList = styled.ul`
-  padding-left: 20px;
-  list-style: disc;
-
-  & > li {
-    margin-bottom: 0.5em;
-
-    &:last-child {
-      margin-bottom: 0;
-    }
-  }
-`;
